+++ src/__tests__/config.test.ts
@@ -38,6 +38,7 @@ describe('config.ts', () => {
         "configMap": undefined,
         "githubToken": "n/a",
         "isPullRequest": false,
+        "manualPlugins": Array [],
         "options": Object {
           "color": undefined,
           "diff": undefined,
@@ -103,6 +104,7 @@ describe('config.ts', () => {
         "configMap": undefined,
         "githubToken": "n/a",
         "isPullRequest": false,
+        "manualPlugins": Array [],
         "options": Object {
           "color": undefined,
           "diff": undefined,
@@ -160,6 +162,7 @@ describe('config.ts', () => {
         "configMap": undefined,
         "githubToken": "n/a",
         "isPullRequest": true,
+        "manualPlugins": Array [],
         "options": Object {
           "color": undefined,
           "diff": undefined,